Consolidated route controllers
